Removed redundant photo draft dao (refactored photo dao/room repo), draft now created before opening form (id generated), on exit draft save prompt/auto drop if empty, draft auto-saved again w/ delay after change, price now always stored in usd in db, other various fixes, currently implementing form validation/errors

+++ app/src/main/java/com/bakjoul/realestatemanager/data/property/model/PropertyDto.kt
@@ -13,7 +13,7 @@ data class PropertyDto(
     @ColumnInfo("for_sale_since") val forSaleSince: LocalDate,
     @ColumnInfo("date_of_sale") val dateOfSale: LocalDate?,
     val price: BigDecimal,
-    val surface: Double,
+    val surface: BigDecimal,
     val rooms: Int,
     val bathrooms: Int,
     val bedrooms: Int,
